# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recall API — personal video memory search engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Incorrect credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List the user's chats, most recently active first
      responses:
        "200":
          description: Chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Start a new chat thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a chat with its full message history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Rename a chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat and its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /chats/{id}/messages:
    post:
      operationId: sendChatMessage
      tags: [chats]
      summary: Send a message and get the assistant's researched reply
      description: >-
        Runs the full search pipeline (context rewrite, intent classification,
        retrieval, person confirmation, reranking, answering) against the
        user's archive. With the default JSON Accept header the response is
        the completed turn. When the request's Accept header includes
        text/event-stream, the server streams Server-Sent Events instead:
        `stage` events ({"stage": "..."}) as each pipeline phase starts,
        then one final `result` event whose data is the same ChatTurn JSON,
        or an `error` event ({"error": "..."}).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Completed turn (user message + assistant reply)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatTurn"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "503":
          description: Search backend not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, name, createdAt, languageProfile]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        createdAt: { type: string }
        languageProfile:
          type: array
          items: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        languageProfile:
          type: array
          items: { type: string }
          description: Languages the user regularly films in, used to flag likely transcription misdetections.
    SignupInput:
      type: object
      required: [email, password, name]
      properties:
        # format: email intentionally omitted — orval emits zod-v4 `zod.email()`
        # against our zod v3 import. The server validates the format instead.
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
        name: { type: string, minLength: 1 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
...
    SearchResult:
      type: object
      required: [id, videoId, videoTitle, thumbnailUrl, snippet, matchType, timestampSeconds, durationSeconds]
      properties:
        id: { type: integer }
        videoId: { type: integer }
        videoTitle: { type: string }
        thumbnailUrl: { type: string }
        videoUrl: { type: ["string", "null"] }
        snippet: { type: string }
        matchType: { type: string, enum: [speech, scene, person, title] }
        matchReason:
          type: ["string", "null"]
          description: One-sentence LLM explanation of why this result matches the query (null when reranking was unavailable)
        timestampSeconds: { type: integer }
        durationSeconds: { type: integer }
        people:
          type: array
          items: { type: string }
        recordedAt: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
...
    PersonFilter:
      type: object
      required: [personName, sceneQuery, status]
      properties:
        personName:
          type: string
          description: Enrolled person detected in the query
        sceneQuery:
          type: string
          description: The scene description actually used for semantic search
        status:
          type: string
          enum: [applied, unavailable]
          description: >-
            "applied" when results were filtered by confirmed face matches;
            "unavailable" when face recognition was unreachable/misconfigured
            and scene-only results are shown instead.
...
    Chat:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ChatDetail:
      allOf:
        - $ref: "#/components/schemas/Chat"
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [id, chatId, role, content, createdAt]
      properties:
        id: { type: integer }
        chatId: { type: integer }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string }
        voice:
          type: boolean
          description: User message originated from voice input
        personIds:
          type: array
          items: { type: integer }
          description: Explicit person mentions attached to a user message
        personNames:
          type: array
          items: { type: string }
          description: Names for the personIds at send time (for display)
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
          description: Clip results backing an assistant reply
        personFilter:
          oneOf:
            - $ref: "#/components/schemas/PersonFilter"
            - type: "null"
        intent:
          type: string
          enum: [search, count, recency, group]
        answer:
          type: ["string", "null"]
        failed:
          type: boolean
          description: Assistant turn ended in an error; content holds the message
    ChatInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 120
    ChatUpdate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 2000
          description: May be empty only when personIds are provided
        personIds:
          type: array
          items: { type: integer }
          maxItems: 5
          description: Enrolled person ids mentioned via "/" pills; multiple ids must ALL appear (AND)
        voice:
          type: boolean
          description: Message text came from voice transcription
    ChatTurn:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        assistantMessage:
          $ref: "#/components/schemas/ChatMessage"
...